+++ internal/manifest/modulecr/client.go
@@ -2,7 +2,6 @@ package modulecr
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -10,7 +9,6 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/kyma-project/lifecycle-manager/api/shared"
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
@@ -18,8 +16,6 @@ import (
 	"github.com/kyma-project/lifecycle-manager/internal/util/collections"
 	"github.com/kyma-project/lifecycle-manager/pkg/util"
 )
-
-var ErrRequeueRequired = errors.New("requeue required")
 
 type Client struct {
 	client.Client
@@ -71,10 +67,18 @@ func (c *Client) CheckCRDeletion(ctx context.Context, manifestCR *v1beta2.Manife
 	return resourceCR == nil, nil
 }
 
+// RemoveModuleCR deletes the module CR if available in the cluster.
+// It uses DeletePropagationBackground to delete module CR.
+// Only if module CR is not found (indicated by NotFound error), it continues to remove Manifest finalizer,
+// and we consider the CR removal successful.
 func (c *Client) RemoveModuleCR(ctx context.Context, kcp client.Client, manifest *v1beta2.Manifest) error {
 	if !manifest.GetDeletionTimestamp().IsZero() {
-		if err := c.deleteCR(ctx, kcp, manifest); err != nil {
+		if err := c.deleteCR(ctx, manifest); err != nil {
 			// we do not set a status here since it will be deleting if timestamp is set.
+			manifest.SetStatus(manifest.GetStatus().WithErr(err))
+			return err
+		}
+		if err := finalizer.RemoveCRFinalizer(ctx, kcp, manifest); err != nil {
 			manifest.SetStatus(manifest.GetStatus().WithErr(err))
 			return err
 		}
@@ -83,13 +87,7 @@ func (c *Client) RemoveModuleCR(ctx context.Context, kcp client.Client, manifest
 	return nil
 }
 
-// deleteCR deletes the module CR if available in the cluster.
-// It uses DeletePropagationBackground to delete module CR.
-// Only if module CR is not found (indicated by NotFound error), it continues to remove Manifest finalizer,
-// and we consider the CR removal successful.
-func (c *Client) deleteCR(
-	ctx context.Context, kcp client.Client, manifest *v1beta2.Manifest,
-) error {
+func (c *Client) deleteCR(ctx context.Context, manifest *v1beta2.Manifest) error {
 	if manifest.Spec.Resource == nil {
 		return nil
 	}
@@ -98,26 +96,10 @@ func (c *Client) deleteCR(
 	propagation := apimetav1.DeletePropagationBackground
 	err := c.Delete(ctx, resource, &client.DeleteOptions{PropagationPolicy: &propagation})
 
-	if !util.IsNotFound(err) {
-		return nil
-	}
-
-	onCluster := manifest.DeepCopy()
-	err = kcp.Get(ctx, client.ObjectKeyFromObject(manifest), onCluster)
-	if util.IsNotFound(err) {
-		return fmt.Errorf("deleteCR: %w", err)
-	}
-	if err != nil {
+	if err != nil && !util.IsNotFound(err) {
 		return fmt.Errorf("failed to fetch resource: %w", err)
 	}
-	if removed := controllerutil.RemoveFinalizer(onCluster, finalizer.CustomResourceManagerFinalizer); removed {
-		if err := kcp.Update(
-			ctx, onCluster, client.FieldOwner(finalizer.CustomResourceManagerFinalizer),
-		); err != nil {
-			return fmt.Errorf("failed to update resource: %w", err)
-		}
-		return ErrRequeueRequired
-	}
+
 	return nil
 }
